requirements.txt
Django==2.2.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
djangorestframework==3.9.4
gunicorn==19.9.0
psycopg2-binary==2.8.2
requests==2.22.0
